Updated `fedora_submit.py` to use `python3`. I uninstalled the `python2-fedora` package, installed the `python3-fedora` package (and it's automatically-pulled dependencies) on Fedora 28, ran the default regressions, and the submission script worked as expected.
I also tweaked `documentation.txt` to be in Markdown format, plus made
a few minor style changes. The overwhelming majority of the text is
unchanged and the structure is largely the same. The mapping between
the original style and the Markdown version is readily apparent when
the two are viewed side-by-side.